+++ src/internal/DotStarColorFeatures.h
@@ -41,3 +41,10 @@ License along with NeoPixel.  If not, see
 #include "features/DotStarLrgbFeature.h"
 #include "features/DotStarRbgFeature.h"
 #include "features/DotStarRgbFeature.h"
+#include "features/Lpd8806BrgFeature.h"
+#include "features/Lpd8806GrbFeature.h"
+#include "features/Lpd6803BrgFeature.h"
+#include "features/Lpd6803GrbFeature.h"
+#include "features/Lpd6803GbrFeature.h"
+#include "features/Lpd6803RgbFeature.h"
+#include "features/P9813BgrFeature.h"
