+++ src/internal/features/Lpd6803GbrFeature.h
@@ -1,0 +1,73 @@
+/*-------------------------------------------------------------------------
+Lpd6803GbrFeature provides feature class to describe color order and
+color depth for NeoPixelBus template class when used with DotStar like chips
+
+Written by Michael C. Miller.
+
+I invest time and resources providing this open source code,
+please support me by dontating (see https://github.com/Makuna/NeoPixelBus)
+
+-------------------------------------------------------------------------
+This file is part of the Makuna/NeoPixelBus library.
+
+NeoPixelBus is free software: you can redistribute it and/or modify
+it under the terms of the GNU Lesser General Public License as
+published by the Free Software Foundation, either version 3 of
+the License, or (at your option) any later version.
+
+NeoPixelBus is distributed in the hope that it will be useful,
+but WITHOUT ANY WARRANTY; without even the implied warranty of
+MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+GNU Lesser General Public License for more details.
+
+You should have received a copy of the GNU Lesser General Public
+License along with NeoPixel.  If not, see
+<http://www.gnu.org/licenses/>.
+-------------------------------------------------------------------------*/
+#pragma once
+
+
+class  Lpd6803GbrFeature : public Neo2ByteElementsNoSettings
+{
+public:
+    static void applyPixelColor(uint8_t* pPixels, uint16_t indexPixel, ColorObject color)
+    {
+        uint8_t* p = getPixelAddress(pPixels, indexPixel);
+        uint16_t color555;
+
+        encodePixel(color.G, color.B, color.R, &color555);
+        *p++ = color555 >> 8;
+        *p = color555 & 0xff;
+    }
+
+    static ColorObject retrievePixelColor(const uint8_t* pPixels, uint16_t indexPixel)
+    {
+        ColorObject color;
+        const uint8_t* p = getPixelAddress(pPixels, indexPixel);
+
+        uint16_t color555;
+
+        color555 = ((*p++) << 8);
+        color555 |= (*p);
+
+        decodePixel(color555, &color.G, &color.B, &color.R);
+
+        return color;
+    }
+
+    static ColorObject retrievePixelColor_P(PGM_VOID_P pPixels, uint16_t indexPixel)
+    {
+        ColorObject color;
+        const uint8_t* p = getPixelAddress((const uint8_t*)pPixels, indexPixel);
+
+        uint16_t color555;
+
+        color555 = (pgm_read_byte(p++) << 8);
+        color555 |= pgm_read_byte(p);
+
+        decodePixel(color555, &color.G, &color.B, &color.R);
+
+        return color;
+    }
+};
+
